Compiled and executed on 64 bits on Windows. Modified the master CMakeLists.txt file, to use dependencies folder as source for libraries, when present.

+++ source/Core/Castor3D/Src/BufferDeclaration.cpp
@@ -21,7 +21,7 @@ namespace Castor3D
 	}
 
 	BufferDeclaration::BufferDeclaration( std::vector< BufferElementDeclaration > const & p_elements )
-		: BufferDeclaration( p_elements.data(), p_elements.size() )
+		: BufferDeclaration( p_elements.data(), uint32_t( p_elements.size() ) )
 	{
 	}
