+++ source/Core/Castor3D/Src/LightManager.cpp
@@ -86,7 +86,7 @@ namespace Castor3D
 
 			for ( auto && l_it : m_typeSortedLights )
 			{
-				l_lightsCount->GetValue( 0 )[l_it.first] += l_it.second.size();
+				l_lightsCount->GetValue( 0 )[l_it.first] += uint32_t( l_it.second.size() );
 
 				for ( auto l_light : l_it.second )
 				{
@@ -111,7 +111,7 @@ namespace Castor3D
 
 			for ( auto && l_it : m_typeSortedLights )
 			{
-				l_lightsCount->GetValue( 0 )[l_it.first] -= l_it.second.size();
+				l_lightsCount->GetValue( 0 )[l_it.first] -= uint32_t( l_it.second.size() );
 			}
 		}
 	}
